restic: extract UidGidInt into separate files
Extract UidGidInt to simplify moving the locking code to the repository.
